pyproject.toml
[tool.poetry]
name = "fast-api-app"
version = "0.1.0"
description = ""
authors = ["zxc_322 <kvasoliaeww@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
psycopg2-binary = "^2.9.4"
redis = "^4.3.4"
databases = {extras = ["postgresql"], version = "^0.6.1"}
alembic = "^1.8.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^1.10.2"}
bcrypt = "^4.0.1"
fastapi-pagination = "^0.10.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
